Remove ctap-types dependency
We only need the ctap-types for six error codes.  It makes more sense to
define these error codes directly inside the crate so that we don’t need
to make usbd-ctaphid releases just to bump ctap-types.

Arguably, this also makes the code more correct:  Previously, we used a
type that represents the CTAP2 error codes, but we actually want to use
the CTAPHID error codes.  Those happen to be the same, but it could be a
cause for confusion.

In the future, we could also use the ctaphid-types crate instead.