Update to ctaphid-dispatch v0.3.0 and remove MESSAGE_SIZE constant
ctaphid-dispatch v0.3.0 makes Dispatch generic over the message size
instead of using a constant value.  This patch adapts usbd-ctaphid to
this patch and changes the usbd-ctaphid buffer to use the same buffer
size as ctaphid-dispatch.  (Previously, we used 3072 and
ctaphid-dispatch used 7609.)  This means that we no longer need the
MESSAGE_SIZE constant and can remove it.